refactor: extract co-located helpers to clear 1000-line file gate
The main merge tipped ChannelScreen.tsx, ChannelPane.tsx, and AppShell.tsx over the hard 1000-line limit in check-file-sizes.mjs. Moved pure module-level helpers out, byte-identical in behavior: AppShell route/window-drag/notification helpers to AppShell.helpers.ts; ChannelPane channel-intro/system-message helpers to ChannelPane.helpers.ts; ChannelScreen's reply-graph builders to the existing subtreeCreatedAt.ts. The cleared read-state reconcile is untouched.
