+++ desktop/src/features/channels/ui/ChannelPane.helpers.ts
@@ -1,0 +1,53 @@
+import { isEphemeralChannel } from "@/features/channels/lib/ephemeralChannel";
+import type { TimelineMessage } from "@/features/messages/types";
+import type { Channel } from "@/shared/api/types";
+import { KIND_SYSTEM_MESSAGE } from "@/shared/constants/kinds";
+
+export function getChannelIntroKind(channel: Channel): string {
+  const isPrivate = channel.visibility === "private";
+  const isEphemeral = isEphemeralChannel(channel);
+
+  if (isPrivate && isEphemeral) {
+    return "private ephemeral channel";
+  }
+  if (isPrivate) {
+    return "private channel";
+  }
+  if (isEphemeral) {
+    return "ephemeral channel";
+  }
+  return "regular channel";
+}
+
+export function getChannelIntroDescription(channel: Channel): string | null {
+  return (
+    channel.topic?.trim() ||
+    channel.purpose?.trim() ||
+    channel.description.trim() ||
+    null
+  );
+}
+
+export function isWelcomeSetupSystemMessage(message: TimelineMessage) {
+  if (message.kind !== KIND_SYSTEM_MESSAGE) {
+    return false;
+  }
+
+  try {
+    const payload = JSON.parse(message.body) as { type?: string };
+    return (
+      payload.type === "channel_created" || payload.type === "member_joined"
+    );
+  } catch {
+    return false;
+  }
+}
+
+export function mentionsKnownAgent(
+  mentionPubkeys: string[],
+  knownAgentPubkeys: ReadonlySet<string>,
+) {
+  return mentionPubkeys.some((pubkey) =>
+    knownAgentPubkeys.has(pubkey.toLowerCase()),
+  );
+}
